docs: use stardoc for reference documentation only
Previously, extend.md and repository.md were generated with Stardoc.
These files are explanations rather than references, and the doc strings
are in files that weren't meant to be imported. Newer versions of
Stardoc add load statements for the internal .bzl files that we don't
want.

Instead:

- Added reference.md, generated from def.bzl. This contains reference
  documentation for all importable symbols.
- Removed extend.bzl. The explanation here is useful, but we can
  edit extend.md directly and refer to symbols in reference.md.
- Removed repositories.bzl and repositories.md. This only contained
  reference documentation for go_repository without a separate
  explanation, so we can use that instead.

No change to extensions.bzl or extensions.md, which provide reference
documentation for go_deps.

Fixes #763