refactor: update bazel target for go_features
This commit updates the bazel taget for go_features to `@org_golang_google_protobuf//types/gofeaturespb`.
It removes the extra `protobuf`.
Updated the template to remove an extra empty line.
Run the `gen_known_imports` to reflect the changes in `proto.csv`, this updates the target for
known types from `com_google_protobuf` to `protobuf`.

Signed-off-by: Ismail Gjevori <isgjevori@gmail.com>