language/go: set visibility correctly for nested internal directories
Added pathtools.LastIndex to accomplish this: we should be looking for
the last occurrence of "internal" in the path, not the first.

Added a test for Index in addition to LastIndex and changed the semantics.
Previously, Index ran path.Clean on both strings, but this made the
result meaningless in cases where it changed the strings.

Fixes #2186