walk: fix race condition
Previously in Walk2, after visiting directories listed on the command line
but before visiting RelsToIndex, we set walker.mode to UpdateDirsMode
to avoid further recursion. This write raced with a read of walker.mode
in walker.shouldVisit when called by populateCache.

This change removes the walker.mode field and instead passes the mode
as an explicit parameter to functions that need it.

Fixes #2235