feature: generate `pgoprofile` attribute if default profile detected
Golang has supported profile-guided optimizations during compilation since `v1.20`. When using `go build`, the default value of `-pgo` is `auto`, which will cause Go to perform profile-guided optimization if a pprof file named `default.pgo` is co-located in the `main` package directory. See [Go's Profile-Guided Optimization documentation](https://go.dev/doc/pgo) for more details.

This PR adds a quick detection of `default.pgo`, and if it exists, automatically includes it in the `go_binary` `pgoprofile` attribute. This is a very simple addition to allow easy generation of the default-case for profile-guided optimization.